+++ lib/dry/validation/extensions/hints.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "dry/monads/result"
-
 module Dry
   module Validation
     # Hints extension
@@ -46,7 +44,7 @@ module Dry
         #
         # @api public
         def messages(new_options = EMPTY_HASH)
-          errors.with(hints.to_a, options.merge(**new_options))
+          errors.with(hints(new_options).to_a, options.merge(**new_options))
         end
 
         # Return hint messages